Тесты вынесены в проект TestsForHomeExercises
