+++ cs/Tests2/ObjectComparison.cs
@@ -1,7 +1,11 @@
-﻿using FluentAssertions;
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using FluentAssertions;
 using NUnit.Framework;
+using HomeExercises;
 
-namespace HomeExercises
+namespace TestsForHomeExercises
 {
     public class ObjectComparison
     {
@@ -51,35 +55,6 @@ namespace HomeExercises
                 && actual.Height == expected.Height
                 && actual.Weight == expected.Weight
                 && AreEqual(actual.Parent, expected.Parent);
-        }
-    }
-
-    public class TsarRegistry
-    {
-        public static Person GetCurrentTsar()
-        {
-            return new Person(
-                "Ivan IV The Terrible", 54, 170, 70,
-                new Person("Vasili III of Russia", 28, 170, 60, null));
-        }
-    }
-
-    public class Person
-    {
-        public static int IdCounter = 0;
-        public int Age, Height, Weight;
-        public string Name;
-        public Person Parent;
-        public int Id;
-
-        public Person(string name, int age, int height, int weight, Person parent)
-        {
-            Id = IdCounter++;
-            Name = name;
-            Age = age;
-            Height = height;
-            Weight = weight;
-            Parent = parent;
         }
     }
 }
